added authorization checks

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require('express-validator');
 const Post = require('../models/post');
 const fs = require('fs');
 const path = require('path');
+const User = require('../models/user.js');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -44,20 +45,32 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
+
         imageUrl: imageUrl,
-        creator: {
-            name: 'Stancho'
-        }
+        creator: req.userId
     });
     post.save()
         .then(result => {
+           return User.findById(req.userId)
+        })
+         .then(user => { 
+             creator = user;
+             user.posts.push(post); 
+             return user.save(); 
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfuly!',
-                post: result
-            })
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
+            });
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -70,6 +83,7 @@ exports.createPost = (req, res, next) => {
 exports.getPost = (req, res, next) => {
     const postId = req.params.postId;
     Post.findById(postId)
+        .populate('creator')
         .then(post => {
             if (!post) {
                 const error = new Error('Could not find post');
@@ -112,6 +126,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -140,11 +159,22 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({ message: 'Post deleted' })
         })
         .catch(err => {
